refactor: extract cli launcher
The launcher command is not tied to the main function and can be used
for testing.
