+++ src/lib.rs
@@ -5,6 +5,7 @@
 //! The central concept of Monotax is an income. An income is a record of money received by a person.
 
 pub mod app;
+pub mod cli;
 pub mod config;
 pub mod db;
 pub mod domain;
